cryptpandas/requirements.txt
pandas
cryptography
pyarrow
cryptpandas
azure-keyvault-secrets==4.6.0
azure-identity==1.12.0
pytest
pytest-cov
